Server: WIP-structure proejct setings

+++ server/db/seeds.rb
@@ -1,5 +1,8 @@
 # coding: utf-8
-Project.destroy_all
+if Project.any?
+  Project.destroy_all
+  p "Deleted exisitng projects"
+end
 
 Project.create!(
   name: "Blog",
@@ -23,16 +26,6 @@ Project.create!(
   active_database: "default"
 )
 
-project = Project.create!(
-  name: "Adventure",
-  description: "Write an adventure story and make your friends face-to-face",
-  public: true,
-  slug: "cyoa",
-  preview: "9a253ae0-8af1-4056-8984-50cb37b79c55",
-  index_page_id: "352150eb-88bf-451b-821e-9fed8ce02cc2",
-  active_database: "default"
-)
-
 Project.create!(
   name: "Test: Sequence DB",
   description: "Dieses Projekt wird für automatische Tests benutzt und hat keine inhaltliche  Bedeutung.",
@@ -42,10 +35,20 @@ Project.create!(
   active_database: "default"
 )
 
+# build project with an association to project sources
+project = Project.create!(
+  name: "Adventure",
+  description: "Write an adventure story and make your friends face-to-face",
+  public: true,
+  slug: "cyoa",
+  preview: "9a253ae0-8af1-4056-8984-50cb37b79c55",
+  index_page_id: "352150eb-88bf-451b-821e-9fed8ce02cc2",
+  active_database: "default"
+)
 project.project_sources.create!(
   url: "https://de.wikipedia.org/wiki/Spielbuch",
   title: "Wikipedia",
   display: "The sample story comes 1: 1 from the Wikipedia article to Spielbuch"
 )
 
-p "Created #{Project.count} projects"
+p "Created #{Project.count} projects and #{ProjectSource.count} resources for #{project.slug}"
